Routes extracted into separate module, #112
