+++ src/simod_http/.gitignore
@@ -1,1 +1,2 @@
 tmp
+data
